+++ src/main/java/com/yuntun/sanitationkitchen/service/impl/TrashWeightSerialServiceImpl.java
@@ -32,12 +32,12 @@ public class TrashWeightSerialServiceImpl extends ServiceImpl<TrashWeightSerialM
   }
 
   @Override
-  public Integer countCurrentWeight(TrashWeightSerialDto dto) {
-    Integer total = weightSerialMapper.countCurrentWeight(dto);
+  public Double countCurrentWeight(TrashWeightSerialDto dto) {
+    Double total = weightSerialMapper.countCurrentWeight(dto);
     if (total != null) {
       return total;
     } else {
-      return 0;
+      return 0D;
     }
   }
